Add files via upload
Take a look into some of my designs for a online food ordering app.

Onboarding Screen:-This screen serves as users' introduction to our app's features and benefits. 
The Onboarding Screen is the first interaction users have with our app. It's designed to provide a quick glimpse into the app's functionality, showcase its value proposition, and encourage users to get started.

Thank You Screen:-This screen aims to leave a positive and appreciative final impression on users.The Thank You Screen is a crucial touchpoint in the user journey. It's where we express gratitude for users' orders and assure them of our commitment to their satisfaction.

Item Selection Screen:-This screen enhances users' selection experience and guides them through their choices.The Item Selection Screen is a core part of the app's functionality, enabling users to browse and choose their desired items seamlessly.